functional and styled

+++ public/scripts/client.js
@@ -36,20 +36,25 @@ function displayTasks(tasks){
       $tableRow.append($checkBox);
     }
 
-    $tableRow.append('<td><input type="text" name="task" value="' + task.task +'"/></td>');
-    $tableRow.append('<td><input type="text" name="notes" value="' + task.notes + '"/></td>');
+    $tableRow.append('<td><input class="focus editFields" type="text" name="task" value="' + task.task +'"/></td>');
+    $tableRow.append('<td><input class="focus editFields" type="text" name="notes" value="' + task.notes + '"/></td>');
 
     // var $checkBox = $(('<td><input id="checkIt" type="checkbox" name="complete" value="'+task.complete+'"/></td>'));
     // $checkBox.data('id', task.id);
     // $tableRow.append($checkBox);
+    var $now = $.now();
+    var time = $now.toLocaleDateString;
+    if (task.update == null) {
+      $tableRow.append('<td id="theDate">' + $now + '</td>');
+    } else {
+      $tableRow.append('<td id="theDate">' + task.updated + '</td>');
+    }
 
-    //$tableRow.append('<td id="theDate">' + task.updated + '</td>');
-
-    var $updateButton = $('<td><button class="update btn btn-outline-success btn-sm">Update</button></td>');
+    var $updateButton = $('<td><button class="update btn btn-outline-success btn-sm"><i class="fa fa-pencil" aria-hidden="true"></i></button></td>');
     $updateButton.data('id', task.id);
     $tableRow.append($updateButton);
 
-    var $deleteButton = $('<td><button class="delete btn btn-outline-danger btn-sm">Delete</button></td>');
+    var $deleteButton = $('<td><button class="delete btn btn-outline-danger btn-sm"><i class="fa fa-trash" aria-hidden="true"></i></button></td>');
     $deleteButton.data('id', task.id);
     $tableRow.append($deleteButton);
 
@@ -62,7 +67,8 @@ function displayTasks(tasks){
     //   $("#checkIt").attr('checked', true);
     // };
     //$('#checkIt:checked').parent().siblings().children().css('text-decoration', 'line-through');
-    $("[value=false]").parent().siblings().children().css({'text-decoration': 'line-through', 'color': 'lightgray'});
+    // $("[value=false]").parent().siblings().children().css({'text-decoration': 'line-through', 'color': 'lightgray'});
+    $("[value=false]").parent().siblings().find('input').css({'text-decoration': 'line-through', 'color': 'lightgray'});
 
   });
 }
@@ -115,10 +121,13 @@ function updateTask(event) {
 }
 
 function deleteTask (event){
-  event.preventDefault();
-  $.ajax({
-    url: '/tasks/' + $(this).parent().data('id'),
-    type: 'DELETE',
-    success: getTasks
-  });
+  var confirmation = confirm ('Are you sure you want to delete this entry?');
+  if (confirmation == true) {
+    event.preventDefault();
+    $.ajax({
+      url: '/tasks/' + $(this).parent().data('id'),
+      type: 'DELETE',
+      success: getTasks
+    });
+  }
 }
